Complete section 16. [Atoms]
Atoms are used for `Uncoordinated synchronous` access to a single
identity. Use when just one thing needs to be changed by multiple
threads.

Values stored as atoms are kept up to date but using a retry loop
whenever changes are applied to them.

They do not need to run in transactions.

because of the retry loop, functions that change the atom must be free
of side effects (because the function is run multiple times.)